refactor(acl): split ApplicationController into 3 single-action controllers
Each controller handles exactly one HTTP action (SOLID / SRP):

- ApplyToRoleController    → POST   /acl/{role_id}/apply          (acl.apply)
- ApproveApplicationController → POST /acl/{role_id}/approve/{user_id} (acl.approve)
- DenyApplicationController    → DELETE /acl/{role_id}/deny/{user_id}  (acl.deny)

Delete ApplicationController.php.

Update routes/Routes/AccessControl/View.php.

Split ApplicationControllerTest into 3 matching test files:
- ApplyToRoleControllerTest        (5 tests: unauth, on-request, 403 x3)
- ApproveApplicationControllerTest (3 tests: unauth, non-mod 403, approve)
- DenyApplicationControllerTest    (3 tests: unauth, non-mod 403, deny)

256 tests pass.
